+++ src/app/dashboard/escrow/initialize-escrow/page.tsx
@@ -1,8 +1,8 @@
 "use client";
 
-import CreateEscrow from "@/components/modules/escrow/CreateEscrow";
+import CreateEscrow from "@/components/modules/escrow/ui/pages/CreateEscrow";
 import { EscrowSteps } from "@/components/modules/escrow/ui/steps/EscrowStepper";
-import InitEscrow from "@/components/modules/escrow/InitEscrow";
+import InitEscrow from "@/components/modules/escrow/ui/pages/InitEscrow";
 
 const InitializeEscrowPage: React.FC = () => {
   const steps = [
