+++ src/components/modules/escrow/ui/pages/InitEscrow.tsx
@@ -3,8 +3,8 @@
 import FlipCard from "@/components/utils/Code/FlipCard";
 import Loader from "@/components/utils/Loader";
 import { useLoaderStore } from "@/store/utilsStore/store";
-import { initializeEscrowCode } from "./code/initialize.code";
-import InitializeEscrowForm from "./ui/forms/InitializeEscrowForm";
+import { initializeEscrowCode } from "../../code/initialize.code";
+import InitializeEscrowForm from "../forms/InitializeEscrowForm";
 
 const InitializeEscrow = () => {
   const isLoading = useLoaderStore((state) => state.isLoading);
